requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
requests=2.31.0